failless-docs-bot/requirements.txt
requests
python-telegram-bot
pillow
